requirements.txt
PyQt5>=5.15
PyOpenGL>=3.1.5
numpy>=1.21
scipy>=1.7.0
